Reintegrate `TracerWebMvcAutoConfiguration` into main configuration class as per review comment
No suitable testing scenario for this setup is available,
because TracerAutoConfiguration implicitly imports
TracerWebMvcAutoConfiguration.
